RDISCROWD-4839: Refactor app_settings, settings_local (#674)
* refactor app_settings, settings_local

* refactor run.py, hdfs->pybhdfs

* fix test

* launch settings_test from pybossa module; same as settings_local

* update settings_upref_mdata loc
