+++ .github/workflows/github-actions.yml
@@ -40,7 +40,7 @@ jobs:
           redis-server --version
           redis-server contrib/redis/sentinel.conf --sentinel
           cd pybossa && ln -s themes/default/translations && cd ..
-          cp settings_test.py.tmpl settings_test.py
+          cp settings_test.py.tmpl pybossa/settings_test.py
 
       - name: Install dependencies
         run: |
