+++ pybossa/app_settings.py
