+++ pybossa/settings_local.py.tmpl
