+++ run_tests
@@ -1,2 +1,2 @@
 #!/bin/bash
-PYBOSSA_SETTINGS="./settings_test.py" PYBOSSA_REDIS_CACHE_DISABLED=1 nosetests ${@:-test/}
+PYBOSSA_SETTINGS="./pybossa/settings_test.py" PYBOSSA_REDIS_CACHE_DISABLED=1 nosetests ${@:-test/}
